add changes to the comment
