+++ src/style.css
@@ -121,7 +121,9 @@ form{
 
 .name-field{
   display: flex;
-  gap: 20px;
+  flex-direction: column;
+  gap: 10px;
+  margin-top: 10px;
 }
 
 .popup-container {
@@ -196,4 +198,15 @@ form{
 .close-icon:hover {
   color: red;
   transform: scale(1.5);
+}
+
+li{
+  list-style: none;
+  gap: 12px;
+  font-size: 10px;
+}
+
+.form-container{
+  display: flex;
+  gap: 10px;
 }
